Move `CalculateReward` method from RuneHelper to WorldbossHelper
