+++ Lib9c/Battle/RaidSimulatorV1.cs
@@ -173,7 +173,7 @@ namespace Nekoyume.Battle
             }
 
             var rank =  WorldBossHelper.CalculateRank(_currentBossRow, DamageDealt);
-            AssetReward = RuneHelper.CalculateReward(
+            AssetReward = WorldBossHelper.CalculateReward(
                 rank,
                 BossId,
                 _runeWeightSheet,
