+++ Lib9c/Helper/WorldBossHelper.cs
@@ -1,5 +1,9 @@
 using System;
+using System.Collections.Generic;
+using System.Linq;
+using Libplanet.Action;
 using Libplanet.Types.Assets;
+using Nekoyume.Battle;
 using Nekoyume.Model.State;
 using Nekoyume.TableData;
 
@@ -43,6 +47,67 @@ namespace Nekoyume.Helper
             return refillInterval > 0 &&
                    (blockIndex - startedIndex) / refillInterval >
                    (refilledIndex - startedIndex) / refillInterval;
+        }
+
+        public static List<FungibleAssetValue> CalculateReward(
+            int rank,
+            int bossId,
+            RuneWeightSheet sheet,
+            IWorldBossRewardSheet rewardSheet,
+            RuneSheet runeSheet,
+            IRandom random
+        )
+        {
+            var row = sheet.Values.First(r => r.Rank == rank && r.BossId == bossId);
+            var rewardRow =
+                rewardSheet.OrderedRows.First(r => r.Rank == rank && r.BossId == bossId);
+            if (rewardRow is WorldBossKillRewardSheet.Row kr)
+            {
+                kr.SetRune(random);
+            }
+            else if (rewardRow is WorldBossBattleRewardSheet.Row rr)
+            {
+                rr.SetRune(random);
+            }
+
+            var total = 0;
+            var dictionary = new Dictionary<int, int>();
+            while (total < rewardRow.Rune)
+            {
+                var selector = new WeightedSelector<int>(random);
+                foreach (var info in row.RuneInfos)
+                {
+                    selector.Add(info.RuneId, info.Weight);
+                }
+
+                var ids = selector.Select(1);
+                foreach (var id in ids)
+                {
+                    if (dictionary.ContainsKey(id))
+                    {
+                        dictionary[id] += 1;
+                    }
+                    else
+                    {
+                        dictionary[id] = 1;
+                    }
+                }
+
+                total++;
+            }
+
+#pragma warning disable LAA1002
+            var result = dictionary
+#pragma warning restore LAA1002
+                .Select(kv => RuneHelper.ToFungibleAssetValue(runeSheet[kv.Key], kv.Value))
+                .ToList();
+
+            if (rewardRow.Crystal > 0)
+            {
+                result.Add(rewardRow.Crystal * CrystalCalculator.CRYSTAL);
+            }
+
+            return result;
         }
     }
 }
